[REGULATION SAVER] Unit tests: moved general part to separated context
